+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+- Use Chooser for MIDI channel and add global chan.
+- Add categories for encoders and pads
 
 ## [v0.1.1] - 2019-04-16
 ### Changed